clippy::unnecessary_cast
warning: casting to the same type is unnecessary (`usize` -> `usize`)
   --> src/sniffer.rs:289:24
    |
289 |                 chains[chr as usize].add_observation(freq);
    |                        ^^^^^^^^^^^^ help: try: `chr`
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
    = note: `#[warn(clippy::unnecessary_cast)]` on by default

+++ src/sniffer.rs
@@ -286,7 +286,7 @@ impl Sniffer {
                 }
             }
             for (chr, &freq) in freqs.iter().enumerate() {
-                chains[chr as usize].add_observation(freq);
+                chains[chr].add_observation(freq);
             }
         }
